Extract worker utility methods to separate file.
